PureInput Hotfix
I'm assuming InnerValue was to safely control the component's bound value without exposing the parent value.

Trouble is, we never properly updated this.

Rather than fix that, I opted to abuse the fact that we wrapped the bind in the first place, since TextValue only needs to be read-only and OnChange/OnInput pass it up the parent bind